Add Makefile to install um on Linux
The gem is built and then installed as the current user, on the
expectation that I'll be installing it under a version of Ruby that's
installed within the current user's home directory.

The completion script is installed into /etc/bash_completion.d, which is
the correct location on both Ubuntu and Solus. The man pages are copied
into /usr/local.

Both of these last two installations are carried out as root, via sudo,
so there's no need to run `make` itself with sudo. If you do, you'll
find yourself installing um with your system Ruby (which may be fine,
but isn't what I've decided to do right now).